User Summary VC updated from Patron Info screen.
